requirements.txt
openai
uagents
uagents-core
python-dotenv
requests
